Remove `DiagramPlotter`, differentiate between implementations of `plot`
Also make plot methods static in classes in `gtda.homology` and in ForgetDimensino
