+++ gtda/homology/cubical.py
@@ -10,11 +10,12 @@ from sklearn.utils.validation import check_array, check_is_fitted
 
 from ._utils import _pad_diagram
 from ..externals.python import CubicalComplex, PeriodicCubicalComplex
+from ..plotting import plot_diagram
 from ..utils.intervals import Interval
 from ..utils.validation import validate_params
 
 
-class CubicalPersistence(BaseEstimator, TransformerMixin):
+class CubicalPersistence(BaseEstimator, TransformerMixin, PlotterMixin):
     """:ref:`Persistence diagrams <persistence diagram>` resulting from
     :ref:`filtered cubical complexes <cubical complex>`.
 
@@ -220,3 +221,25 @@ class CubicalPersistence(BaseEstimator, TransformerMixin):
         Xt = np.stack(Xt)
         Xt = np.nan_to_num(Xt, posinf=self.infinity_values_)
         return Xt
+
+    @staticmethod
+    def plot(Xt, sample=0, homology_dimensions=None):
+        """Plot a persistence diagram from a collection, with homology in
+        multiple dimensions.
+
+        Parameters
+        ----------
+        Xt : ndarray of shape (n_samples, n_points, 3)
+            Collection of persistence diagrams, such as returned by
+            :meth:`transform`.
+
+        sample : int, optional, default: ``0``
+            Index of the sample in `Xt` to be plotted.
+
+        homology_dimensions : list, tuple or None, optional, default: ``None``
+            Which homology dimensions to include in the plot. ``None`` means
+            plotting all dimensions present in ``Xt[sample]``.
+
+        """
+        return plot_diagram(
+            Xt[sample], homology_dimensions=homology_dimensions)
